# docs/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bank API",
    "description": "Документация REST API для Bank API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Authorization"
        ],
        "summary": "Регистрация пользователя",
        "description": "Позволяет зарегистироваться пользователю",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Пользователь успешно зарегистрирован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации данных запроса",
            "content": {
              "application/json": {

              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {

              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Authorization"
        ],
        "summary": "Обновление токена",
        "description": "Позволяет обновить access токен",
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Authorization"
        ],
        "summary": "Выход пользователя",
        "description": "Позволяет осуществить выход пользователя",
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "Успешный выход",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authorization"
        ],
        "summary": "Вход пользователя",
        "description": "Позволяет осуществить вход пользователю",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Успешный вход",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации данных запроса",
            "content": {
              "application/json": {

              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {

              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "login": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ROLE_USER",
              "ROLE_ADMIN"
            ]
          }
        }
      },
      "RegisterRequestDTO": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "maxLength": 2147483647,
            "minLength": 8,
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).*$"
          },
          "rememberMe": {
            "type": "boolean"
          }
        },
        "required": [
          "login",
          "password"
        ]
      },
      "MessageResponseDTO": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "LoginRequestDTO": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          },
          "rememberMe": {
            "type": "boolean"
          }
        },
        "required": [
          "login",
          "password"
        ]
      }
    }
  }
}